WeatherInfoFragment classes moved to weatherInfoFragment package
