+++ app/src/main/java/com/simplertutorials/android/wheathograophy/components/ApplicationComponent.kt
@@ -2,7 +2,7 @@ package com.simplertutorials.android.wheathograophy.components
 
 import javax.inject.Singleton
 import com.simplertutorials.android.wheathograophy.ui.fragments.cityListFragment.CityListFragment
-import com.simplertutorials.android.wheathograophy.ui.fragments.WeatherInfoFragment
+import com.simplertutorials.android.wheathograophy.ui.fragments.weatherInfoFragment.WeatherInfoFragment
 import dagger.Component
 
 @Singleton
