+++ app/src/main/java/com/simplertutorials/android/wheathograophy/ui/MainActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import androidx.fragment.app.Fragment
 import com.simplertutorials.android.wheathograophy.R
 import com.simplertutorials.android.wheathograophy.ui.fragments.cityListFragment.CityListFragment
-import com.simplertutorials.android.wheathograophy.ui.fragments.WeatherInfoFragment
+import com.simplertutorials.android.wheathograophy.ui.fragments.weatherInfoFragment.WeatherInfoFragment
 import com.simplertutorials.android.wheathograophy.ui.fragments.addCityFragment.AddCityFragment
 
 class MainActivity : AppCompatActivity() {
